requirements.txt
av==11.0.0
blinker==1.7.0
certifi==2024.2.2
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
coloredlogs==15.0.1
ctranslate2==4.1.0
faster-whisper==1.0.1
filelock==3.13.3
Flask==3.0.3
flatbuffers==24.3.25
fsspec==2024.3.1
huggingface-hub==0.22.2
humanfriendly==10.0
idna==3.6
importlib_metadata==7.1.0
itsdangerous==2.1.2
Jinja2==3.1.3
lxml==5.2.1
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
onnxruntime==1.17.1
packaging==24.0
pillow==10.3.0
protobuf==5.26.1
pysrt==1.1.2
python-docx==1.1.0
pyuwsgi==2.0.23.post0
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
safetensors==0.4.2
sentencepiece==0.2.0
sympy==1.12
tokenizers==0.15.2
torch==2.2.2
torchvision==0.17.2
tqdm==4.66.2
transformers==4.39.3
typing_extensions==4.11.0
urllib3==2.2.1
Werkzeug==3.0.2
wget==3.2
zipp==3.18.1

psycopg2~=2.9.9
python-dotenv~=1.0.1